requirements.txt
flask==3.0.0
flask-cors==4.0.0
langchain==0.1.0
langchain-community==0.0.10
langchain-openai==0.0.2
openai==1.6.1
pymilvus==2.3.4
psycopg2-binary==2.9.9
sentence-transformers==2.2.2
transformers==4.36.2
torch==2.1.2
torchvision==0.16.2
Pillow==10.1.0
PyMuPDF==1.23.14
pandas==2.1.4
numpy==1.24.4
python-dotenv==1.0.0
streamlit==1.29.0
chromadb==0.4.18
faiss-cpu==1.7.4
requests==2.31.0
beautifulsoup4==4.12.2
python-multipart==0.0.6
uvicorn==0.25.0
fastapi==0.108.0
pydantic==2.5.2
sqlalchemy==2.0.23
alembic==1.13.1
redis==5.0.1
celery==5.3.4
gunicorn==21.2.0